refactor: 플레이어의 재생바 컴포넌트 외부 파일로 분리
